+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.1.1 - 2020-10-11
+### Fixed
+- Issue with explained list where not all items were outputted when a similarly
+named key existed.
+
 ## 1.1.0 - 2020-05-02
 ### Added
 - Success styling.
@@ -26,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created the initial implementation of the CLI package.
 
 - [Unreleased](https://github.com/ulrack/cli/compare/1.1.0...HEAD)
+- [1.1.0 > 1.1.1](https://github.com/ulrack/cli/compare/1.1.0...1.1.1)
 - [1.0.3 > 1.1.0](https://github.com/ulrack/cli/compare/1.0.3...1.1.0)
 - [1.0.2 > 1.0.3](https://github.com/ulrack/cli/compare/1.0.2...1.0.3)
 - [1.0.1 > 1.0.2](https://github.com/ulrack/cli/compare/1.0.1...1.0.2)